pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cosmic-stars-prototype"
version = "0.1.0"
description = "Solar system N-body backend prototype based on REBOUND/REBOUNDx."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [
  { name = "Cosmic Stars Team" }
]
dependencies = [
  "colour-science",
  "numpy",
  "OpenEXR",
  "pillow",
  "rebound",
  "reboundx",
]

[project.scripts]
cosmic-sim = "main_simulation:main"
cosmic-integrator-check = "integrator_check:main"
cosmic-bake-blackbody-lut = "blackbody_1d_lut_baker:main"
cosmic-bake-far-field-lut = "far_field_lut_baker:main"
cosmic-exr-preview = "exr_preview:main"

[tool.setuptools]
package-dir = { "" = "src" }
py-modules = ["blackbody_1d_lut_baker", "database", "exr_preview", "extra_function", "far_field_lut_baker", "integrator_check", "main_simulation"]
